move ix start stuff into `DBoxInterface`
