+++ pets-lib/dg/src/main.dg
@@ -1,7 +1,6 @@
-% Main
 ###
 
-Import ./rodrick.dg
+Import rodrick.dg
 
 ###
 ---
