+++ pets-lib/src/lib.rs
@@ -33,7 +33,9 @@ mod prelude {
     pub use crate::macros::*;
     pub use crate::stats::*;
 
-    pub use crate::dialogue::ix_list;
+    pub use crate::dialogue::autoload::DBoxInterface;
+    pub use crate::dialogue::ix_map;
+
     pub use crate::world::interaction::manager::InteractionManager;
 
     // is this bad practice? no clue and idc honestly
